requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
web3==6.11.1
pandas==2.1.3
python-dotenv==1.0.0
gunicorn==21.2.0 